+++ test/test_model.py
@@ -82,9 +82,8 @@ def test_objective() -> None:
     assert m.objectiverange.min() == 2
     assert m.objectiverange.max() == 2
 
-    # test objective with constant which is not supported
-    with pytest.raises(ValueError):
-        m.objective = m.objective + 3
+    # test objective with constant which is supported
+    m.objective = m.objective + 3
 
 
 def test_remove_variable() -> None: